nf_loto_workspace_ws4_patched/legacy/nf_loto_feature_project/requirements.txt
streamlit
pandas
numpy
psycopg2-binary
sqlalchemy
neuralforecast
tsfresh
tsfel
tsflex
antropy
nolds
stumpy
torch
ray
scikit-learn
matplotlib
jupyter
nbformat
